Add files via upload
Team, this is the working test Hacker News scraper I made.  I wanted to get it out so you can fetch some data, but it is still a work in progress.  I spent a good amount of time handling errors stemming from HTTP get requests.  We will have no control if a linked article rejects the request to scrape text (the program should handle this).  I added a 3 second timeout for the request to external linked articles.  For each "scraping depth" chosen, you will retrieve 30 articles.  If you want to get the article, change the "search_external_bool" to true.  Let me know if you encounter any bugs!  The variable score_sorted is a dictionary that contains the linked articles contents, the title, votes and link.